moved model inited translate
