+++ translations/en.json
@@ -1,0 +1,8 @@
+{
+  "Add device": "Add device",
+  "Attempts": "Attempts",
+  "Device Monitor": "Device Monitor",
+  "Offline interval (s)": "Offline interval (s)",
+  "Online interval (s)": "Online interval (s)",
+  "Status": "Status"
+}
