+++ translations/ru.json
@@ -1,0 +1,8 @@
+{
+  "Add device": "Добавить устройство",
+  "Attempts": "Попытки",
+  "Device Monitor": "Device Monitor",
+  "Offline interval (s)": "Интервал оффлайн (с)",
+  "Online interval (s)": "Интервал онлайн (с)",
+  "Status": "Статус"
+}
